+++ browser/data-browser/src/views/File/fileTypeUtils.ts
@@ -8,5 +8,17 @@ const supportedApplicationFormats = new Set([
   'application/x-sh',
 ]);
 
+const supportedImageTypes = new Set([
+  'image/png',
+  'image/jpeg',
+  'image/gif',
+  'image/svg+xml',
+  'image/webp',
+  'image/avif',
+]);
+
 export const isTextFile = (mimeType: string): boolean =>
   mimeType?.startsWith('text/') || supportedApplicationFormats.has(mimeType);
+
+export const isImageFile = (mimeType: string): boolean =>
+  supportedImageTypes.has(mimeType);